drivers/led/sled1734x: Refactor for reduced RAM usage
This effectively swaps the PWM buffer and config, decreasing the former's
size by making the latter larger. Now the config contains an entry
for each register in each IC, providing a lookup table that allows to
determine the LED index and color channel that register is used for.
Unused registers should be set to an index of NO_LED (255).

The PWM buffer has been replaced with an LED buffer, storing a 3 byte RGB
structure for each LED, thus minimizing its size from 256 to 3 * (LED
count). In combination with the new config, this allows us to construct
the relevant chunk of the buffer at flush time, without adding too much
complexity or performance overhead (I have not noticed any impact, but
I'm just eyeballing it).

This means that now the buffer size is dependent on the amount of LEDs
defined. On keyboards with less LEDs the savings will be greater.

Previous buffer size:
* 256 bytes for PWM registers
* 32 bytes for control registers
* a few bytes for the "dirty" state booleans
=> ~294 bytes per IC

New buffer size:
* 3 * (LED count) bytes
* 32 bytes for control registers
* a few bytes for the "dirty" state booleans
=> ~38 b per IC + 3 b per LED

On the Redragon Aryaman K569, which uses two ICs and has 118 LEDs,
this translates to:

* previously: 294 bytes * 2 ICs = 588 bytes
* now: ((32 b + 6 b) * 2 ICs) + (3 colors * 118 LEDs) = 430 bytes
=> difference: 588 b - 430 b = 158 bytes saved